refactor: 불필요한 컴포넌트 합치기 및  폼 검증 함수 별도로 빼기
