+++ src/components/pages/Everytime/Everytime.tsx
@@ -1,15 +1,50 @@
 import NavigationButtons from "../../ui/custom/navigationButton";
-import { EverytimeSchema, type EverytimeValues } from "./Everytime.Schema";
+import type { EverytimeValues } from "./Everytime.Schema";
 import { LoadingState } from "@/types/state/loading";
-import EverytimeItems from "./Everytime.FieldGroup";
-import { useState, useCallback } from "react";
-import { ZodError } from "zod";
+import EverytimeTimeTableLink from "./Everytime.TimeTableLink";
+import { useState, useCallback, useEffect } from "react";
+import { Button } from "@/components/ui/button";
+import { CheckCircleIcon, LoaderCircle } from "lucide-react";
+import { ClockAlert } from "lucide-react";
+import AlertBox from "../../ui/custom/alertbox";
+import type React from "react";
+import validateEverytime from "./Everytime.Validate";
 
 interface EverytimeProps {
   onNext: () => void;
   onPrev: () => void;
-  onDataSubmit: (data: EverytimeValues) => void; // 데이터 제출 핸들러
+  onDataSubmit: (data: EverytimeValues) => void;
 }
+
+interface StatusMessageProps {
+  loading: LoadingState;
+}
+
+const StatusMessage = ({ loading }: StatusMessageProps) => {
+  switch (loading) {
+    case LoadingState.LOADING:
+      return (
+        <>
+          <LoaderCircle
+            className="h-8 w-8 animate-spin text-gray-500"
+            style={{ animation: "spin 3s linear infinite" }}
+          />
+          <p className="pl-4">시간표 정보를 읽는 중 입니다. . .</p>
+        </>
+      );
+
+    case LoadingState.SUCCESS:
+      return (
+        <>
+          <CheckCircleIcon className="h-8 w-8 text-green-400" />
+          <p className="pl-4 text-green-400">제출이 완료되었습니다 !</p>
+        </>
+      );
+
+    default:
+      return null;
+  }
+};
 
 function Everytime({ onNext, onPrev, onDataSubmit }: EverytimeProps) {
   // 폼 상태 관리
@@ -30,31 +65,14 @@ function Everytime({ onNext, onPrev, onDataSubmit }: EverytimeProps) {
     setEverytimeValues((prev) => ({ ...prev, [name]: value }));
   }, []);
 
-  // 폼 유효성 검사 함수
-  const validate = useCallback((data: EverytimeValues) => {
-    try {
-      EverytimeSchema.parse(data); // 데이터 검증
-      setError({}); // 기존 에러 초기화
-      return true; // 유효성 검사 통과
-    } catch (error) {
-      // 유효성 검사 실패 시 에러 처리
-      if (error instanceof ZodError) {
-        const newError: { timetableLink?: { message?: string } } = {};
-        for (const err of error.errors) {
-          if (err.path[0] === "timetableLink") {
-            newError.timetableLink = { message: err.message };
-          }
-        }
-        setError(newError);
-      }
-      return false; // 유효성 검사 실패
-    }
-  }, []);
-
   // 폼 제출 함수 (콜백으로 메모이제이션)
   const handleSubmit = useCallback(async () => {
-    if (!validate(everytimeValues)) {
+    // 유효성 검사 함수 호출
+    const validationResult = validateEverytime(everytimeValues);
+
+    if (!validationResult.success) {
       // 유효성 검사 실패 시 제출 중단
+      setError(validationResult.error || {});
       setLoading(LoadingState.IDLE); // 로딩 상태 초기화
       return;
     }
@@ -65,20 +83,58 @@ function Everytime({ onNext, onPrev, onDataSubmit }: EverytimeProps) {
     await new Promise((resolve) => setTimeout(resolve, 3000)); // await 추가
     setLoading(LoadingState.SUCCESS); // 로딩 상태 성공으로 변경
     onDataSubmit(everytimeValues); // 데이터 제출
-  }, [everytimeValues, validate, onDataSubmit]);
+  }, [everytimeValues, onDataSubmit]);
+
+  // onNext 함수 래핑
+  const handleNext = () => {
+    const validationResult = validateEverytime(everytimeValues);
+    if (validationResult.success) {
+      setError({}); // 유효성 검사 성공 시 에러 상태 초기화
+      onNext();
+    } else {
+      setError(validationResult.error || {}); // 에러 상태 업데이트
+    }
+  };
+
+  useEffect(() => {
+    const validationResult = validateEverytime(everytimeValues);
+    if (validationResult.success) {
+      setError({}); // 유효성 검사 성공 시 에러 상태 초기화
+    }
+  }, [everytimeValues]);
 
   return (
     <div className="mb-12 space-y-4">
-      <EverytimeItems
-        everytimeValues={everytimeValues}
-        error={error}
-        loading={loading}
-        onSubmit={handleSubmit}
-        onChange={handleChange}
+      <h3 className="font-semibold text-lg">에브리타임 시간표 제출</h3>
+      <AlertBox
+        icon={<ClockAlert className="h-4 w-4" />}
+        title="시간표 제출이 왜 필요한가요?"
+        description={[
+          "활동을 계획할 때 수업과 겹치지 않게 계획하기 위해서 시간표가 필요해요.",
+        ]}
       />
+      <form
+        className="my-10 space-y-2"
+        onSubmit={(e) => {
+          e.preventDefault();
+          handleSubmit(); // onSubmit 함수 호출
+        }}
+      >
+        <EverytimeTimeTableLink
+          timetableLink={everytimeValues.timetableLink}
+          onChange={handleChange} // onChange 함수 전달
+          error={error}
+        />
+        <div className="mt-4 flex items-center space-x-4">
+          <Button className="mt-2 inline" type="submit" disabled={loading === LoadingState.LOADING}>
+            {loading === LoadingState.LOADING ? "제출 중..." : "제출"}
+          </Button>
+          <StatusMessage loading={loading} />
+        </div>
+      </form>
       <NavigationButtons
         prev={onPrev}
-        next={onNext} // 다음 버튼 클릭 시 onNext 호출
+        next={handleNext} // handleNext 함수 전달
         isValid={Object.keys(error).length === 0 && loading === LoadingState.SUCCESS}
       />
     </div>
